refactor(Login, Register): remove default values from form fields
In both the Login and Register components, the default values for the form fields have been removed. This change ensures that users must enter their information explicitly, enhancing the user experience by preventing pre-filled forms that might not apply to every user. This update affects the email and password fields in the Login component, as well as the name, lastName, location, email, and password fields in the Register component.