fix(server): update static file path
Update the static file path in server.js to point to the correct directory in the client/dist folder instead of the public folder. This change ensures that the correct index.html file is served when accessing routes that do not match any defined routes.

feat(package): update dependencies

- Updated @babel/runtime to version 7.24.7 with new integrity and resolved URLs.
- Added @mongodb-js/saslprep as an optional dependency with version 1.1.7.
- Added sparse-bitfield as a dependency.
- Updated @types/node to version 20.14.8 with undici-types updated to ~5.26.4.
- Updated @types/webidl-conversions to version 7.0.3 with a new integrity URL.
- Removed the package abbrev version 1.1.1.
- Updated axios to version 1.7.2 with new integrity and resolved URLs.

feat: add setup-production-app script

Add a new script in package.json to set up the production app by installing dependencies, building the client, and setting up the server. This script can be run with 'npm run setup-production-app'.